refactor: separate config loading
